WIP, refactored files, tests pass :) TODOs: * Add tests for more complex SQL queries * Fill out unfinished tests
